Quick hack to make EmulationStation handle the GP Wiz40 arcade control board.
This required adjusting the DEADZONE value and changing the way movement around the DEADZONE is filtered.

The root cause is due to the control board using values 0, 127, and 255 to indicate movements(full negative, centered, full positive).

A better fix would be to have EmulationStation auto calibrate the axes or provide some configuration options.